# Optimize `labelValuesWithMatchers` to fetch label values from selected series