--- a/tsdb/querier.go
+++ b/tsdb/querier.go
@@ -346,6 +346,11 @@ func inversePostingsForMatcher(ix IndexPostingsReader, m *labels.Matcher) (index
 	return ix.Postings(m.Name, res...)
 }
 
+const (
+	minSeriesForPostingsLabelValues = 1000 // Minimum number of label values when we try to use labels from series instead.
+	maxExpandedFactor               = 100  // Division factor for maximum number of matched series.
+)
+
 func labelValuesWithMatchers(r IndexReader, name string, matchers ...*labels.Matcher) ([]string, error) {
 	p, err := PostingsForMatchers(r, matchers...)
 	if err != nil {
@@ -355,6 +360,36 @@ func labelValuesWithMatchers(r IndexReader, name string, matchers ...*labels.Mat
 	allValues, err := r.LabelValues(name)
 	if err != nil {
 		return nil, errors.Wrapf(err, "fetching values of label %s", name)
+	}
+
+	// If there are many values, let's see if expanded postings for matchers have smaller cardinality.
+	if len(allValues) >= minSeriesForPostingsLabelValues {
+		// We will only expand this number of postings.
+		maxPostings := len(allValues) / maxExpandedFactor
+		// Add space for one extra posting when checking if we expanded all.
+		expanded := make([]storage.SeriesRef, 0, maxPostings+1)
+
+		for len(expanded) < maxPostings && p.Next() {
+			expanded = append(expanded, p.At())
+		}
+
+		// Did we reach end of postings?
+		reachedEnd := true
+		if len(expanded) == maxPostings && p.Next() {
+			reachedEnd = false
+			expanded = append(expanded, p.At())
+		}
+
+		if reachedEnd {
+			// We have expanded all the postings. Given that number of postings is
+			// much smaller than number of all label values, let's compute label
+			// values from series instead.
+			// (If we did this for all postings, it would be very expensive.)
+			return labelValuesFromSeries(r, name, expanded)
+		}
+
+		// If we haven't reached end of postings, we prepend our expanded postings to "p", and continue.
+		p = newPrependPostings(expanded, p)
 	}
 
 	// If we have a matcher for the label name, we can filter out values that don't match
@@ -394,6 +429,78 @@ func labelValuesWithMatchers(r IndexReader, name string, matchers ...*labels.Mat
 	}
 
 	return values, nil
+}
+
+// labelValuesFromSeries returns all unique label values from for given label name from supplied series. Values are not sorted.
+func labelValuesFromSeries(r IndexReader, labelName string, refs []storage.SeriesRef) ([]string, error) {
+	values := map[string]struct{}{}
+
+	var builder labels.ScratchBuilder
+	for _, ref := range refs {
+		err := r.Series(ref, &builder, nil)
+		if err != nil {
+			return nil, errors.Wrapf(err, "label values for label %s", labelName)
+		}
+
+		v := builder.Labels().Get(labelName)
+		if v != "" {
+			values[v] = struct{}{}
+		}
+	}
+
+	result := make([]string, 0, len(values))
+	for v := range values {
+		result = append(result, v)
+	}
+	return result, nil
+}
+
+func newPrependPostings(a []storage.SeriesRef, b index.Postings) index.Postings {
+	return &prependPostings{
+		ix:     -1,
+		prefix: a,
+		rest:   b,
+	}
+}
+
+// prependPostings returns series references from "prefix" before using "rest" postings.
+type prependPostings struct {
+	ix     int
+	prefix []storage.SeriesRef
+	rest   index.Postings
+}
+
+func (p *prependPostings) Next() bool {
+	p.ix++
+	if p.ix < len(p.prefix) {
+		return true
+	}
+	return p.rest.Next()
+}
+
+func (p *prependPostings) Seek(v storage.SeriesRef) bool {
+	for p.ix < len(p.prefix) {
+		if p.ix >= 0 && p.prefix[p.ix] >= v {
+			return true
+		}
+		p.ix++
+	}
+
+	return p.rest.Seek(v)
+}
+
+func (p *prependPostings) At() storage.SeriesRef {
+	if p.ix >= 0 && p.ix < len(p.prefix) {
+		return p.prefix[p.ix]
+	}
+	return p.rest.At()
+}
+
+func (p *prependPostings) Err() error {
+	if p.ix >= 0 && p.ix < len(p.prefix) {
+		return nil
+	}
+	return p.rest.Err()
 }
 
 func labelNamesWithMatchers(r IndexReader, matchers ...*labels.Matcher) ([]string, error) {