--- a/tsdb/querier.go
+++ b/tsdb/querier.go
@@ -346,10 +346,7 @@ func inversePostingsForMatcher(ix IndexPostingsReader, m *labels.Matcher) (index
 	return ix.Postings(m.Name, res...)
 }
 
-const (
-	minSeriesForPostingsLabelValues = 1000 // Minimum number of label values when we try to use labels from series instead.
-	maxExpandedFactor               = 100  // Division factor for maximum number of matched series.
-)
+const maxExpandedPostingsFactor = 100 // Division factor for maximum number of matched series.
 
 func labelValuesWithMatchers(r IndexReader, name string, matchers ...*labels.Matcher) ([]string, error) {
 	p, err := PostingsForMatchers(r, matchers...)
@@ -360,36 +357,6 @@ func labelValuesWithMatchers(r IndexReader, name string, matchers ...*labels.Mat
 	allValues, err := r.LabelValues(name)
 	if err != nil {
 		return nil, errors.Wrapf(err, "fetching values of label %s", name)
-	}
-
-	// If there are many values, let's see if expanded postings for matchers have smaller cardinality.
-	if len(allValues) >= minSeriesForPostingsLabelValues {
-		// We will only expand this number of postings.
-		maxPostings := len(allValues) / maxExpandedFactor
-		// Add space for one extra posting when checking if we expanded all.
-		expanded := make([]storage.SeriesRef, 0, maxPostings+1)
-
-		for len(expanded) < maxPostings && p.Next() {
-			expanded = append(expanded, p.At())
-		}
-
-		// Did we reach end of postings?
-		reachedEnd := true
-		if len(expanded) == maxPostings && p.Next() {
-			reachedEnd = false
-			expanded = append(expanded, p.At())
-		}
-
-		if reachedEnd {
-			// We have expanded all the postings. Given that number of postings is
-			// much smaller than number of all label values, let's compute label
-			// values from series instead.
-			// (If we did this for all postings, it would be very expensive.)
-			return labelValuesFromSeries(r, name, expanded)
-		}
-
-		// If we haven't reached end of postings, we prepend our expanded postings to "p", and continue.
-		p = newPrependPostings(expanded, p)
 	}
 
 	// If we have a matcher for the label name, we can filter out values that don't match
@@ -409,6 +376,34 @@ func labelValuesWithMatchers(r IndexReader, name string, matchers ...*labels.Mat
 			}
 		}
 		allValues = filteredValues
+	}
+
+	// Let's see if expanded postings for matchers have smaller cardinality than label values.
+	// Since computing label values from series is expensive, we apply a limit on number of expanded
+	// postings (and series).
+	maxExpandedPostings := len(allValues) / maxExpandedPostingsFactor
+	if maxExpandedPostings > 0 {
+		// Add space for one extra posting when checking if we expanded all postings.
+		expanded := make([]storage.SeriesRef, 0, maxExpandedPostings+1)
+
+		for len(expanded) < maxExpandedPostings && p.Next() {
+			expanded = append(expanded, p.At())
+		}
+
+		// Did we reach end of postings?
+		reachedEnd := true
+		if len(expanded) == maxExpandedPostings && p.Next() {
+			reachedEnd = false
+			expanded = append(expanded, p.At())
+		}
+
+		if reachedEnd {
+			// We have expanded all the postings -- all returned label values will be from these series only.
+			return labelValuesFromSeries(r, name, expanded)
+		}
+
+		// If we haven't reached end of postings, we prepend our expanded postings to "p", and continue.
+		p = newPrependPostings(expanded, p)
 	}
 
 	valuesPostings := make([]index.Postings, len(allValues))