This PR optimizes labelValues operation with matchers by iterating over all series matched by provided matchers, and collecting unique label values.

On its own this is pretty expensive operation, so this is only activated when number of matched series is tiny compared to number of all values for given label name.

Specifically, we require that queried label has at least 1000 unique label values, and matchers matched less than `(number of label values / 100)` series. (I did not do any analysis to compute these numbers, and I'm open to suggestions how to come with better ones.)

Benchmark for labelValuesForMatchers (updated for https://github.com/grafana/mimir-prometheus/pull/518/commits/e4d815d2383ba307c24a2cc72d55897d1e58179a):

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
                                                                         │   before.txt   │              after.txt              │
                                                                         │     sec/op     │    sec/op     vs base               │
Querier/Head/labelValuesWithMatchers/i_with_i="1"-10                          1.159m ± 3%   1.142m ±  3%        ~ (p=0.310 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="1"-10                          97.22m ± 4%   95.98m ±  4%        ~ (p=0.485 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="^.+$"-10                       119.8m ± 4%   119.4m ± 17%        ~ (p=0.937 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j!="foo"-10                 113.4m ± 4%   112.2m ±  3%        ~ (p=0.180 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j=~"X.+"-10               18295.8µ ± 6%   908.4µ ±  2%  -95.04% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j=~"XXX|YYY"-10           17927.7µ ± 6%   902.2µ ±  2%  -94.97% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="X",j!="foo"-10                 90.19m ± 5%   83.65m ±  3%   -7.25% (p=0.004 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"-10       126.8m ± 7%   120.4m ±  8%        ~ (p=0.180 n=6)
Querier/Head/labelValuesWithMatchers/i_with_i_times_n=533701-10            375686.4µ ± 7%   893.0µ ±  2%  -99.76% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_i_times_n=20-10                233119.2µ ± 7%   898.4µ ±  4%  -99.61% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_i_times_n=~"12.*""-10             387.8m ± 5%   403.0m ±  3%        ~ (p=0.093 n=6)
Querier/Head/labelValuesWithMatchers/n_with_j!="foo"-10                       248.6m ± 4%   247.8m ±  2%        ~ (p=0.310 n=6)
Querier/Head/labelValuesWithMatchers/n_with_i="1"-10                          5.301µ ± 1%   5.310µ ±  2%        ~ (p=0.937 n=6)
Querier/Block/labelValuesWithMatchers/i_with_i="1"-10                         1.325m ± 1%   1.323m ±  2%        ~ (p=0.818 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="1"-10                         86.92m ± 1%   89.00m ±  1%   +2.39% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="^.+$"-10                      101.2m ± 2%   101.4m ±  1%        ~ (p=0.818 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="1",j!="foo"-10                152.4m ± 2%   159.0m ±  1%   +4.31% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="1",j=~"X.+"-10               45.283m ± 1%   1.193m ±  2%  -97.37% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="1",j=~"XXX|YYY"-10           45.355m ± 1%   1.205m ±  3%  -97.34% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="X",j!="foo"-10                227.7m ± 1%   231.8m ±  1%   +1.80% (p=0.009 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"-10      155.0m ± 1%   162.3m ±  1%   +4.69% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_i_times_n=533701-10            123.369m ± 1%   1.098m ±  1%  -99.11% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_i_times_n=20-10                103.955m ± 1%   1.101m ±  2%  -98.94% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_i_times_n=~"12.*""-10            114.7m ± 0%   119.7m ±  1%   +4.30% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/n_with_j!="foo"-10                      293.9m ± 2%   292.7m ±  1%   -0.41% (p=0.015 n=6)
Querier/Block/labelValuesWithMatchers/n_with_i="1"-10                         2.358m ± 1%   2.358m ±  1%        ~ (p=0.485 n=6)
geomean                                                                       47.22m        12.73m        -73.05%

                                                                         │  before.txt   │               after.txt               │
                                                                         │     B/op      │     B/op      vs base                 │
Querier/Head/labelValuesWithMatchers/i_with_i="1"-10                        1.531Mi ± 0%   1.531Mi ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/labelValuesWithMatchers/i_with_n="1"-10                        16.14Mi ± 0%   16.15Mi ± 0%   +0.05% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="^.+$"-10                     32.31Mi ± 0%   32.32Mi ± 0%   +0.02% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j!="foo"-10               16.14Mi ± 0%   16.15Mi ± 0%   +0.05% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j=~"X.+"-10              10.695Mi ± 0%   1.539Mi ± 0%  -85.61% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j=~"XXX|YYY"-10          10.695Mi ± 0%   1.539Mi ± 0%  -85.61% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="X",j!="foo"-10               16.14Mi ± 0%   16.15Mi ± 0%   +0.05% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"-10     17.67Mi ± 0%   17.68Mi ± 0%   +0.04% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_i_times_n=533701-10            10.695Mi ± 0%   1.539Mi ± 0%  -85.61% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_i_times_n=20-10                10.695Mi ± 0%   1.540Mi ± 0%  -85.60% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_i_times_n=~"12.*""-10           19.89Mi ± 0%   19.90Mi ± 0%   +0.04% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/n_with_j!="foo"-10                     5.532Ki ± 0%   5.532Ki ± 0%        ~ (p=1.000 n=6)
Querier/Head/labelValuesWithMatchers/n_with_i="1"-10                        4.445Ki ± 0%   4.445Ki ± 0%        ~ (p=1.000 n=6) ¹
Querier/Block/labelValuesWithMatchers/i_with_i="1"-10                       1.531Mi ± 0%   1.531Mi ± 0%        ~ (p=1.000 n=6) ¹
Querier/Block/labelValuesWithMatchers/i_with_n="1"-10                       17.66Mi ± 0%   17.67Mi ± 0%   +0.04% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="^.+$"-10                    33.84Mi ± 0%   33.85Mi ± 0%   +0.02% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="1",j!="foo"-10              17.66Mi ± 0%   17.67Mi ± 0%   +0.04% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="1",j=~"X.+"-10             12.222Mi ± 0%   1.540Mi ± 0%  -87.40% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="1",j=~"XXX|YYY"-10         12.221Mi ± 0%   1.539Mi ± 0%  -87.40% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="X",j!="foo"-10              17.66Mi ± 0%   17.67Mi ± 0%   +0.04% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"-10    19.19Mi ± 0%   19.20Mi ± 0%   +0.04% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_i_times_n=533701-10           12.221Mi ± 0%   1.540Mi ± 0%  -87.40% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_i_times_n=20-10               12.221Mi ± 0%   1.540Mi ± 0%  -87.40% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_i_times_n=~"12.*""-10          21.41Mi ± 0%   21.42Mi ± 0%   +0.04% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/n_with_j!="foo"-10                    7.078Ki ± 0%   7.078Ki ± 0%        ~ (p=1.000 n=6) ¹
Querier/Block/labelValuesWithMatchers/n_with_i="1"-10                       5.977Ki ± 0%   5.977Ki ± 0%        ~ (p=1.000 n=6) ¹
geomean                                                                     3.902Mi        2.106Mi       -46.03%
¹ all samples are equal

                                                                         │   before.txt    │               after.txt               │
                                                                         │    allocs/op    │  allocs/op   vs base                  │
Querier/Head/labelValuesWithMatchers/i_with_i="1"-10                            5.000 ± 0%    5.000 ± 0%         ~ (p=1.000 n=6) ¹
Querier/Head/labelValuesWithMatchers/i_with_n="1"-10                           200.0k ± 0%   200.0k ± 0%    +0.00% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="^.+$"-10                        300.1k ± 0%   300.1k ± 0%    +0.00% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j!="foo"-10                  200.0k ± 0%   200.0k ± 0%    +0.00% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j=~"X.+"-10              200010.000 ± 0%    9.000 ± 0%  -100.00% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j=~"XXX|YYY"-10           200012.00 ± 0%    11.00 ± 0%   -99.99% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="X",j!="foo"-10                  200.0k ± 0%   200.0k ± 0%    +0.00% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"-10        200.0k ± 0%   200.0k ± 0%    +0.00% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_i_times_n=533701-10            200009.000 ± 0%    8.000 ± 0%  -100.00% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_i_times_n=20-10                 200011.00 ± 0%    11.00 ± 0%   -99.99% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/i_with_i_times_n=~"12.*""-10              208.9k ± 0%   208.9k ± 0%    +0.00% (p=0.002 n=6)
Querier/Head/labelValuesWithMatchers/n_with_j!="foo"-10                         98.00 ± 0%    98.00 ± 0%         ~ (p=1.000 n=6) ¹
Querier/Head/labelValuesWithMatchers/n_with_i="1"-10                            87.00 ± 0%    87.00 ± 0%         ~ (p=1.000 n=6) ¹
Querier/Block/labelValuesWithMatchers/i_with_i="1"-10                           7.000 ± 0%    7.000 ± 0%         ~ (p=1.000 n=6) ¹
Querier/Block/labelValuesWithMatchers/i_with_n="1"-10                          300.0k ± 0%   300.0k ± 0%    +0.00% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="^.+$"-10                       400.1k ± 0%   400.1k ± 0%    +0.00% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="1",j!="foo"-10                 300.0k ± 0%   300.0k ± 0%    +0.00% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="1",j=~"X.+"-10              300013.00 ± 0%    12.00 ± 0%  -100.00% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="1",j=~"XXX|YYY"-10          300012.00 ± 0%    11.00 ± 0%  -100.00% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="X",j!="foo"-10                 300.0k ± 0%   300.0k ± 0%    +0.00% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"-10       300.0k ± 0%   300.0k ± 0%    +0.00% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_i_times_n=533701-10            300011.00 ± 0%    16.00 ± 0%   -99.99% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_i_times_n=20-10                300013.00 ± 0%    31.00 ± 0%   -99.99% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/i_with_i_times_n=~"12.*""-10             308.9k ± 0%   308.9k ± 0%    +0.00% (p=0.002 n=6)
Querier/Block/labelValuesWithMatchers/n_with_j!="foo"-10                        141.0 ± 0%    141.0 ± 0%         ~ (p=1.000 n=6) ¹
Querier/Block/labelValuesWithMatchers/n_with_i="1"-10                           129.0 ± 0%    129.0 ± 0%         ~ (p=1.000 n=6) ¹
geomean                                                                        34.09k        1.627k        -95.23%
¹ all samples are equal
```