@pstibrany 2023-07-26
Thanks for review @pracucci . I've updated the PR (and benchmarks) after addressing your comments.