@samjewell 2023-07-27
Just wanted to say thank you @pstibrany (and also thanks to @pracucci and @colega). This looks perfect for my use-case ❤️ 
cc @gtk-grafana